requirements.txt
huawei_lte_api==1.10.0
python-dotenv==1.0.1
